+++ src/RestrictedBoltzmannMachines.jl
@@ -43,6 +43,7 @@ include("ais.jl")
 
 include("train/infinite_minibatches.jl")
 include("train/initialization.jl")
+include("train/train.jl")
 include("train/pcd.jl")
 include("train/ucd.jl")
 include("train/gradient.jl")